Refactor logic

+++ backend/src/main/java/ru/medmentor/dto/SimulationStatsOverviewDto.java
@@ -1,13 +1,11 @@
 package ru.medmentor.dto;
 
-import java.math.BigDecimal;
-
 public record SimulationStatsOverviewDto(
         long completedSessions,
-        BigDecimal averagePoliteness,
-        BigDecimal averageQuestioningStructure,
-        BigDecimal averageThoroughness,
-        BigDecimal averageEmpathy,
-        BigDecimal averageDiagnosisCorrect
+        Double averagePoliteness,
+        Double averageQuestioningStructure,
+        Double averageThoroughness,
+        Double averageEmpathy,
+        Double averageDiagnosisCorrect
 ) {
 }
